refactor: the stack graduates out of dismiss.ts into topmost.ts
architecture-first-principles, C3 (the graduation sweep) and its scope rule.
dismiss.ts held two independent things: the two GESTURES that shut a panel
(and what counts as inside them), and the ORDER that says which panel a
gesture is for. They change on different axes — a third gesture would touch
one, a modal that blocks everything below would touch the other — and the
second already has a consumer that wants none of the first: menu/Dropdown.tsx
takes only the order, since its gestures are Kobalte's already.

One consumer that needs only half is what graduates a half, which is the same
argument arriving.ts was extracted on one directory along. No behaviour
changes; the sweep in claims.test.ts and the docs follow the module.
